Dividing status_code by 100 gives a floating-point number in Python 3; we always want an integer.

+++ metadata_layer.py
@@ -793,7 +793,7 @@ class MetaToModelUtility(object):
             )
             return
 
-        if representation.status_code / 100 not in (2,3):
+        if int(representation.status_code / 100) not in (2,3):
             self.log.info(
                 "Representation %s gave %s status code, not mirroring.",
                 representation.url, representation.status_code
